+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Implemented enhancements
 - No ticket - Improve label of client model fields in django admin
 - No ticket - update gov-notify sumbission type in testapi
+- TT-1922 - activity-stream-integration
 - TT-1975 - implement rate-limiting
 
 ### Fixed bugs: